requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
reportlab
PyPDF2==2.12.1
